+++ .gitignore
@@ -1,0 +1,4 @@
+results/
+.vscode/
+__pycache__/
+*.pyc
