+++ plr_exercise.egg-info/SOURCES.txt
@@ -7,3 +7,5 @@ plr_exercise.egg-info/SOURCES.txt
 plr_exercise.egg-info/dependency_links.txt
 plr_exercise.egg-info/requires.txt
 plr_exercise.egg-info/top_level.txt
+plr_exercise/models/__init__.py
+plr_exercise/models/cnn.py
